Converts the DiffPlex project to a PCL.
This targets .NET 4.0 as it did originally, and it targets SL5, so that the SL5 project is no longer required.
It also targets Windows 8.0, Windows Phone 8.0 and Windows Phone Appx 8.1.

So in short, this greatly expands the allowed users of this library and with fewer projects required.
It also works around the NuProj build issue since referencing the SL5 project is no longer required.